Selenium, Genlogin and API integration
