+++ crawl/crawl_with_api.py
@@ -1,161 +1,148 @@
-import random
 import time
+import random
+import requests
 import pandas as pd
 from selenium import webdriver
-from selenium.webdriver.chrome.service import Service 
+from selenium.webdriver.chrome.service import Service
 from selenium.webdriver.common.by import By
-from selenium.webdriver.support.ui import WebDriverWait
-from selenium.webdriver.support import expected_conditions as EC
+from selenium.webdriver.chrome.options import Options
 from selenium.webdriver.common.action_chains import ActionChains
-
 from Genlogin import Genlogin
+import re
 
-class ShopeeCrawler:
-    def __init__(self, profile_id=None):
-        self.driver = None
-        self.gen = Genlogin("")
-        self.profile_id = profile_id  # Cho phép truyền profile_id bên ngoài
+def inject_anti_bot_scripts(driver):
+    # Xóa navigator.webdriver
+    driver.execute_cdp_cmd("Page.addScriptToEvaluateOnNewDocument", {
+        "source": """
+            Object.defineProperty(navigator, 'webdriver', {get: () => undefined});
+        """
+    })
+    # Fake plugin length
+    driver.execute_script("""
+    Object.defineProperty(navigator, 'plugins', { get: () => [1,2,3,4,5] });
+    """)
+    driver.execute_script("""
+    Object.defineProperty(navigator, 'languages', {get: () => ['vi-VN', 'en-US']});
+    """)
 
-    def _human_delay(self, min_delay=1.5, max_delay=4.5):
-        time.sleep(random.uniform(min_delay, max_delay) + random.random())
-
-    def _get_user_agents(self):
-        return [
-            "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/135.0.0.0 Safari/537.36",
-            "Mozilla/5.0 (iPhone; CPU iPhone OS 16_6 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/16.6 Mobile/15E148 Safari/604.1",
-            "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/136.0.0.0 Safari/537.36"
-        ]
-
-    def _start_stealth_browser(self):
+def simulate_human_behavior(driver):
+    for _ in range(random.randint(5, 10)):
+        driver.execute_script(f"window.scrollBy(0, {random.randint(100, 600)})")
+        time.sleep(random.uniform(0.5, 1.8))
         try:
-            # Nếu chưa truyền profile_id, lấy profile đầu tiên
-            if not self.profile_id:
-                profile_data = self.gen.getProfiles(0, 1)
-                if not profile_data.get("profiles"):
-                    raise Exception("Không tìm thấy profile")
-                self.profile_id = profile_data["profiles"][0]["id"]
-            # Chạy profile được chỉ định
-            run_data = self.gen.runProfile(self.profile_id)
-            if not run_data.get("wsEndpoint"):
-                raise Exception("Không thể khởi chạy profile")
-            debugger_address = run_data["wsEndpoint"].replace("ws://", "").split("/")[0]
-            options = webdriver.ChromeOptions()
-            options.add_experimental_option("debuggerAddress", debugger_address)
-            options.add_argument(f"user-agent={random.choice(self._get_user_agents())}")
-            options.add_argument("--disable-blink-features=AutomationControlled")
-            options.add_argument("--disable-infobars")
-            service = Service('crawl/chromedriver.exe')
-            self.driver = webdriver.Chrome(service=service, options=options)
-            self.driver.execute_script("Object.defineProperty(navigator, 'webdriver', {get: () => undefined})")
-            print(f"Khởi động trình duyệt thành công với profile {self.profile_id}")
-        except Exception as e:
-            print(f"Lỗi khởi động trình duyệt: {str(e)}")
-            raise
+            elements = driver.find_elements(By.CSS_SELECTOR, 'a, button')
+            if elements:
+                ActionChains(driver).move_to_element(random.choice(elements)).perform()
+                time.sleep(random.uniform(0.3, 1.2))
+        except: pass
 
-    def _shopee_behavior(self):
-        for _ in range(random.randint(2, 5)):
-            self.driver.execute_script(f"window.scrollBy(0, {random.randint(300, 600)})")
-            self._human_delay(0.6, 1.2)
-            ActionChains(self.driver).move_by_offset(random.randint(10, 100), random.randint(10, 100)).perform()
-            self._human_delay(0.4, 1.1)
-            try:
-                logo = self.driver.find_element(By.CSS_SELECTOR, "a.navbar__logo")
-                ActionChains(self.driver).move_to_element(logo).click().perform()
-            except Exception:
-                pass
+def get_cookies_from_genlogin(profile_id):
+    gen = Genlogin("")
+    run_data = gen.runProfile(profile_id)
+    if not run_data.get("wsEndpoint"):
+        raise Exception("Không thể khởi chạy profile")
+    debugger_address = run_data["wsEndpoint"].replace("ws://", "").split("/")[0]
 
-    def _handle_captcha(self):
-        print("Phát hiện CAPTCHA! Hãy giải Captcha thủ công trên trình duyệt (kéo thanh hoặc click hình), rồi nhấn Enter để tiếp tục...")
-        input("Sau khi giải xong Captcha trên trình duyệt, nhấn Enter để tiếp tục...")
+    options = Options()
+    options.add_experimental_option("debuggerAddress", debugger_address)
+    service = Service('crawl/chromedriver.exe')
+    driver = webdriver.Chrome(service=service, options=options)
+    inject_anti_bot_scripts(driver)
 
-    def _extract_reviews(self):
-        reviews = []
-        page = 1
-        while True:
-            try:
-                WebDriverWait(self.driver, 20).until(
-                    EC.presence_of_element_located((By.CSS_SELECTOR, "div.shopee-product-rating"))
-                )
-                self._human_delay(0.7, 1.5)
-                for el in self.driver.find_elements(By.CSS_SELECTOR, "div.shopee-product-rating"):
-                    try:
-                        content = el.find_element(By.CSS_SELECTOR, ".shopee-product-rating__main").text
-                    except Exception:
-                        content = ""
-                    try:
-                        rating = len(el.find_elements(By.CSS_SELECTOR, ".icon-rating-solid--active"))
-                    except Exception:
-                        rating = None
-                    try:
-                        author = el.find_element(By.CSS_SELECTOR, ".shopee-product-rating__author-name").text
-                    except Exception:
-                        author = ""
-                    try:
-                        date = el.find_element(By.CSS_SELECTOR, ".shopee-product-rating__time").text
-                    except Exception:
-                        date = ""
-                    reviews.append({
-                        "content": content,
-                        "rating": rating,
-                        "author": author,
-                        "date": date,
-                        "page": page
-                    })
-                next_btn = self.driver.find_elements(By.CSS_SELECTOR, ".shopee-icon-button.shopee-icon-button--right")
-                if next_btn and next_btn[0].is_enabled():
-                    next_btn[0].click()
-                    page += 1
-                    self._human_delay(1.2, 2.8)
-                else:
-                    break
-            except Exception as e:
-                print(f" Lỗi trích xuất trang review: {str(e)}")
-                break
-        return reviews
+    driver.get("https://shopee.vn/")
+    print("Đăng nhập shopee và thao tác qua như người thật để tránh bị detect là bot")
+    input("Sau khi login & thao tác đủ, nhấn Enter để tiếp tục...")
+    simulate_human_behavior(driver)
+    cookies = driver.get_cookies()
+    driver.quit()
+    gen.stopProfile(profile_id)
+    cookie_dict = {c['name']: c['value'] for c in cookies}
+    return cookie_dict
 
-    def crawl(self, product_url):
-        try:
-            if not product_url.startswith(("http://", "https://")):
-                product_url = f"https://{product_url}"
-            self._start_stealth_browser()
-            self.driver.get(product_url)
-            self._human_delay(2, 3)
-            # Nếu bị captcha
-            while "verify" in self.driver.current_url or "captcha" in self.driver.title.lower():
-                self._handle_captcha()
-                self.driver.get(product_url)
-                self._human_delay(2, 3)
-            self._shopee_behavior()
-            reviews = self._extract_reviews()
-            return reviews
-        except Exception as e:
-            print(f" Lỗi tổng: {str(e)}")
-            return []
-        finally:
-            if self.driver:
-                self.driver.quit()
-            if self.profile_id:
-                self.gen.stopProfile(self.profile_id)
+def crawl_reviews_api(shop_id, item_id, cookie_dict, max_review=200):
+    headers = {
+        "user-agent": "Mozilla/5.0",
+        "referer": f"https://shopee.vn/product/{shop_id}/{item_id}",
+        "cookie": "; ".join([f"{k}={v}" for k, v in cookie_dict.items()])
+    }
+    all_reviews = []
+    offset = 0
+    limit = 50
+    while len(all_reviews) < max_review:
+        url = f"https://shopee.vn/api/v2/item/get_ratings?itemid={item_id}&shopid={shop_id}&limit={limit}&offset={offset}"
+        resp = requests.get(url, headers=headers)
+        if resp.status_code != 200:
+            print(f"Lỗi hoặc bị block tại offset {offset}: {resp.text[:200]}")
+            break
+        data = resp.json()
+        ratings = data.get("data", {}).get("ratings", [])
+        if not ratings:
+            break
+        all_reviews.extend(ratings)
+        print(f"Đã lấy {len(all_reviews)} review...")
+        offset += limit
+        time.sleep(2 + (offset % 7))  # delay tránh bị block
+    return all_reviews
+
+def extract_simple_reviews(raw_reviews):
+    reviews = []
+    for r in raw_reviews:
+        reviews.append({
+            "rating": r.get("rating_star"),
+            "username": r.get("author_username"),
+            "comment": r.get("comment"),
+            "date": r.get("ctime"),
+            "images": ",".join(r.get("images", [])) if r.get("images") else ""
+        })
+    return reviews
 
 if __name__ == "__main__":
     # Hiển thị danh sách profile để chọn
     gen = Genlogin("")
-    profiles = gen.getProfiles(0, 20).get("profiles", [])
+    profiles = gen.getProfiles(0, 10).get("profiles", [])
     if not profiles:
         print("Không có profile nào trong Genlogin.")
         exit(1)
     print("Danh sách profile:")
     for i, profile in enumerate(profiles):
-        # In ra để biết chắc key nào là tên
         print(f"{i+1}. {profile.get('profileName') or profile.get('title') or profile.get('name') or 'NoName'} (id: {profile.get('id') or profile.get('profile_id')})")
-    idx = int(input("Nhập số thứ tự profile muốn dùng: ")) - 1
+    while True:
+        idx_input = input("Nhập số thứ tự profile muốn sử dụng: ").strip()
+        if idx_input.isdigit() and 1 <= int(idx_input) <= len(profiles):
+            idx = int(idx_input) - 1
+            break
+        else:
+            print("Vui lòng nhập số hợp lệ trong danh sách.")
     profile_id = profiles[idx].get('id') or profiles[idx].get('profile_id')
 
-    product_url = input("Nhập URL sản phẩm Shopee: ").strip()
-    crawler = ShopeeCrawler(profile_id=profile_id)
-    reviews = crawler.crawl(product_url)
-    if reviews:
-        pd.DataFrame(reviews).to_csv("shopee_reviews.csv", index=False)
-        print(f"Crawl thành công {len(reviews)} đánh giá!")
+    # Lấy cookie từ profile Genlogin
+    cookie_dict = get_cookies_from_genlogin(profile_id)
+
+    product_url = input("Nhập link sản phẩm Shopee: ").strip()
+    # xử lý link sản phẩm
+    m = re.search(r'/product/(\d+)/(\d+)', product_url)
+    if m:
+        shop_id, item_id = m.group(1), m.group(2)
     else:
-        print("Không tìm thấy đánh giá nào!")
+        shop_id = input("Nhập shop_id: ").strip()
+        item_id = input("Nhập item_id: ").strip()
+    while True:
+        max_review_input = input("Số lượng review muốn lấy tối đa (mặc định 200): ").strip()
+        if not max_review_input:
+            max_review = 200
+            break
+        if max_review_input.isdigit():
+            max_review = int(max_review_input)
+            break
+        else:
+            print("Vui lòng nhập số nguyên hợp lệ hoặc để trống.")
+    
+    # gọi API
+    raw_reviews = crawl_reviews_api(shop_id, item_id, cookie_dict, max_review=max_review)
+    reviews = extract_simple_reviews(raw_reviews)
+    if reviews:
+        df = pd.DataFrame(reviews)
+        df.to_csv("shopee_reviews.csv", index=False)
+        print(f"Đã lưu {len(reviews)} review vào file shopee_reviews.csv")
+    else:
+        print("Không lấy được review nào hoặc bị block.")
